+++ elmyra.ip.access.epo/CHANGES.rst
@@ -11,6 +11,7 @@ development
     - /country/us
     - /cpc/"H04B7-15542"
 - ux: link to espacenet worldwide (biblio data)
+- ui: add rating widget “jQuery Raty”
 
 0.15.0
 ======