+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,9 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+
+0.16.0
+======
 - ux: new quick- and vanity urls
     - /publicationdate/2013-03-03/2013-04-03
     - /publicationdate/2014W10